+++ include/QBDI/Callback.h
@@ -90,12 +90,14 @@ _QBDI_ENABLE_BITMASK_OPERATORS(VMEvent)
  * Structure describing the current VM state
  */
 typedef struct {
-    VMEvent event;           /*!< The event(s) which triggered the callback (must be checked using a mask: event & BASIC_BLOCK_ENTRY).*/
-    rword basicBlockStart;   /*!< The current basic block start address which can also be the execution transfer destination.*/
-    rword basicBlockEnd;     /*!< The current basic block end address which can also be the execution transfer destination.*/
-    rword sequenceStart;     /*!< The current sequence start address which can also be the execution transfer destination.*/
-    rword sequenceEnd;       /*!< The current sequence end address which can also be the execution transfer destination.*/
-    rword lastSignal;        /*!< Not implemented.*/
+    VMEvent event;              /*!< The event(s) which triggered the callback (must be checked using a mask: event & BASIC_BLOCK_ENTRY).*/
+    rword basicBlockStart;      /*!< The current basic block start address which can also be the execution transfer destination.*/
+    rword basicBlockEnd;        /*!< The current basic block end address which can also be the execution transfer destination.*/
+    rword sequenceStart;        /*!< The current sequence start address which can also be the execution transfer destination.*/
+    rword sequenceEnd;          /*!< The current sequence end address which can also be the execution transfer destination.*/
+    rword returnAddressPtr;     /*!< The address of the return address detected during EXEC_TRANSFER_CALL. */
+    rword returnAddressValue;   /*!< The return address detected during EXEC_TRANSFER_CALL. */
+    rword lastSignal;           /*!< Not implemented.*/
 } VMState;
 
 /*! VM callback function type.